requirements.txt
# ===================================================================
# ALLMÉDIAS PWA - DEPENDÊNCIAS
# ===================================================================
# Framework principal
Django>=6.0.0,<7.0.0

# Storage e Cloud (Wasabi S3)
django-storages[s3]>=1.14.0,<2.0.0
boto3>=1.34.0,<2.0.0

# Processamento de imagens e PDFs
Pillow>=10.0.0,<11.0.0
reportlab>=4.0.0,<5.0.0

# Configurações e ambiente
python-decouple>=3.8,<4.0

# Assets estáticos em produção
whitenoise>=6.6.0,<7.0.0
gunicorn>=22.0.0,<23.0.0

# Banco de dados MySQL
mysqlclient>=2.2.0,<3.0.0

# Desenvolvimento e debug
django-extensions>=3.2.0,<4.0.0

# Opcional: Para APIs futuras (se necessário)
djangorestframework>=3.14.0,<4.0.0
django-cors-headers>=4.3.0,<5.0.0

